+++ packages/leaflet-map-adapter/src/layer-adapters/TileAdapter/TileAdapter.ts
@@ -8,7 +8,7 @@ import { BaseAdapter } from '../BaseAdapter';
 
 export class TileAdapter extends BaseAdapter<TileAdapterOptions, TileLayer>
   implements BaseLayerAdapter {
-  addLayer(options: TileAdapterOptions) {
+  addLayer(options: TileAdapterOptions): TileLayer | undefined {
     if (options) {
       const { url, ...opt } = options;
       if (url) {